requirements.txt
torch
transformers
safetensors
tkinter
pyyaml
